requirements.txt
paho-mqtt==1.6.1
Pillow==10.0.0
tkinter==0.1.0
pyinstaller==5.13.0